authentication in action

+++ src/main/java/org/openhmis/webservice/ClientService.java
@@ -6,6 +6,7 @@ import java.util.List;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
+import javax.ws.rs.HeaderParam;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
@@ -18,6 +19,7 @@ import org.apache.log4j.Logger;
 import org.openhmis.code.ClientNameDataQuality;
 import org.openhmis.dto.ClientDTO;
 import org.openhmis.manager.ClientManager;
+import org.openhmis.util.Authentication;
 
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -36,7 +38,9 @@ public class ClientService {
 	@GET
 	@Path("/")
 	@Produces({MediaType.APPLICATION_JSON})
-	public String getClients() throws JsonProcessingException {
+	public String getClients(@HeaderParam("Authorization") String authorization) throws JsonProcessingException {
+		if(!Authentication.googleAuthenticate(authorization))
+			throw new Error("You are not authorized to access this content");
 		List<ClientDTO> clientDTOs = clientManager.getClients();
 		return om.writeValueAsString(clientDTOs);
 	}
@@ -44,7 +48,9 @@ public class ClientService {
 	@POST
 	@Path("/")
 	@Produces({MediaType.APPLICATION_JSON})
-	public String createClient(String data) throws JsonParseException, JsonMappingException, IOException {
+	public String createClient(@HeaderParam("Authorization") String authorization, String data) throws JsonParseException, JsonMappingException, IOException {
+		if(!Authentication.googleAuthenticate(authorization))
+			throw new Error("You are not authorized to access this content");
 		// This endpoint takes in a raw json STRING as input.
 		// TODO: support the serialization of individual POST parameters
 		ClientDTO inputVO = om.readValue(data, ClientDTO.class);
@@ -55,7 +61,9 @@ public class ClientService {
 	@GET
 	@Path("/{personalId}")
 	@Produces({MediaType.APPLICATION_JSON})
-	public String getClient(@PathParam("personalId") String personalId) throws JsonProcessingException {
+	public String getClient(@HeaderParam("Authorization") String authorization, @PathParam("personalId") String personalId) throws JsonProcessingException {
+		if(!Authentication.googleAuthenticate(authorization))
+			throw new Error("You are not authorized to access this content");
 		ClientDTO clientDTO = clientManager.getClientByPersonalId(personalId);
 		return om.writeValueAsString(clientDTO);
 	}
@@ -63,7 +71,9 @@ public class ClientService {
 	@PUT
 	@Path("/{personalId}")
 	@Produces({MediaType.APPLICATION_JSON})
-	public String updateClient(@PathParam("personalId") String personalId, String data) throws JsonParseException, JsonMappingException, IOException {
+	public String updateClient(@HeaderParam("Authorization") String authorization, @PathParam("personalId") String personalId, String data) throws JsonParseException, JsonMappingException, IOException {
+		if(!Authentication.googleAuthenticate(authorization))
+			throw new Error("You are not authorized to access this content");
 		ClientDTO inputVO = om.readValue(data, ClientDTO.class);
 		inputVO.setPersonalId(personalId);
 		
@@ -74,7 +84,9 @@ public class ClientService {
 	@DELETE
 	@Path("/{personalId}")
 	@Produces({MediaType.APPLICATION_JSON})
-	public String deleteClient(@PathParam("personalId") String personalId) throws JsonParseException, JsonMappingException, IOException {
+	public String deleteClient(@HeaderParam("Authorization") String authorization, @PathParam("personalId") String personalId) throws JsonParseException, JsonMappingException, IOException {
+		if(!Authentication.googleAuthenticate(authorization))
+			throw new Error("You are not authorized to access this content");
 		clientManager.deleteClient(personalId);
 		return "true";
 	}
