+++ src/main/java/org/openhmis/webservice/OrganizationService.java
@@ -6,6 +6,7 @@ import java.util.List;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
+import javax.ws.rs.HeaderParam;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
@@ -25,6 +26,7 @@ import org.openhmis.manager.FunderManager;
 import org.openhmis.manager.InventoryManager;
 import org.openhmis.manager.OrganizationManager;
 import org.openhmis.manager.SiteManager;
+import org.openhmis.util.Authentication;
 
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -41,7 +43,9 @@ public class OrganizationService {
 	@GET
 	@Path("/")
 	@Produces({MediaType.APPLICATION_JSON})
-	public String getOrganizations() throws JsonProcessingException {
+	public String getOrganizations(@HeaderParam("Authorization") String authorization) throws JsonProcessingException {
+		if(!Authentication.googleAuthenticate(authorization))
+			throw new Error("You are not authorized to access this content");
 		List<OrganizationDTO> organizationDTOs = OrganizationManager.getOrganizations();
 		return om.writeValueAsString(organizationDTOs);
 	}
@@ -49,7 +53,9 @@ public class OrganizationService {
 	@POST
 	@Path("/")
 	@Produces({MediaType.APPLICATION_JSON})
-	public String createOrganization(String data) throws JsonParseException, JsonMappingException, IOException {
+	public String createOrganization(@HeaderParam("Authorization") String authorization, String data) throws JsonParseException, JsonMappingException, IOException {
+		if(!Authentication.googleAuthenticate(authorization))
+			throw new Error("You are not authorized to access this content");
 		// This endpoint takes in a raw json STRING as input.
 		// TODO: support the serialization of individual POST parameters
 		OrganizationDTO inputVO = om.readValue(data, OrganizationDTO.class);
@@ -60,7 +66,9 @@ public class OrganizationService {
 	@GET
 	@Path("/{organizationId}")
 	@Produces({MediaType.APPLICATION_JSON})
-	public String getOrganization(@PathParam("organizationId") String organizationId) throws JsonProcessingException {
+	public String getOrganization(@HeaderParam("Authorization") String authorization, @PathParam("organizationId") String organizationId) throws JsonProcessingException {
+		if(!Authentication.googleAuthenticate(authorization))
+			throw new Error("You are not authorized to access this content");
 		OrganizationDTO organizationDTO = OrganizationManager.getOrganizationByOrganizationId(organizationId);
 		return om.writeValueAsString(organizationDTO);
 	}
@@ -68,7 +76,9 @@ public class OrganizationService {
 	@PUT
 	@Path("/{organizationId}")
 	@Produces({MediaType.APPLICATION_JSON})
-	public String updateOrganization(@PathParam("organizationId") String organizationId, String data) throws JsonParseException, JsonMappingException, IOException {
+	public String updateOrganization(@HeaderParam("Authorization") String authorization, @PathParam("organizationId") String organizationId, String data) throws JsonParseException, JsonMappingException, IOException {
+		if(!Authentication.googleAuthenticate(authorization))
+			throw new Error("You are not authorized to access this content");
 		OrganizationDTO inputVO = om.readValue(data, OrganizationDTO.class);
 		inputVO.setOrganizationId(organizationId);
 		
@@ -79,7 +89,9 @@ public class OrganizationService {
 	@DELETE
 	@Path("/{organizationId}")
 	@Produces({MediaType.APPLICATION_JSON})
-	public String deleteOrganization(@PathParam("organizationId") String organizationId) throws JsonParseException, JsonMappingException, IOException {
+	public String deleteOrganization(@HeaderParam("Authorization") String authorization, @PathParam("organizationId") String organizationId) throws JsonParseException, JsonMappingException, IOException {
+		if(!Authentication.googleAuthenticate(authorization))
+			throw new Error("You are not authorized to access this content");
 		OrganizationManager.deleteOrganization(organizationId);
 		return "true";
 	}
